Update/entities

+++ src/main/java/org/una/programmingIII/Assignment_Manager/Service/Implementation/AssignmentServiceImplementation.java
@@ -11,9 +11,11 @@ import org.una.programmingIII.Assignment_Manager.Exception.ElementNotFoundExcept
 import org.una.programmingIII.Assignment_Manager.Mapper.GenericMapper;
 import org.una.programmingIII.Assignment_Manager.Mapper.GenericMapperFactory;
 import org.una.programmingIII.Assignment_Manager.Model.Assignment;
+import org.una.programmingIII.Assignment_Manager.Model.File;
 import org.una.programmingIII.Assignment_Manager.Repository.AssignmentRepository;
 import org.una.programmingIII.Assignment_Manager.Service.AssignmentService;
 
+import java.time.LocalDate;
 import java.util.*;
 import java.util.stream.Collectors;
 
@@ -47,13 +49,23 @@ public class AssignmentServiceImplementation implements AssignmentService {
 
     @Override
     public AssignmentDto findById(Long id) {
-        return assignmentMapper.convertToDTO(assignmentRepository.findById(id).orElseThrow(() -> new ElementNotFoundException("FeedingSchedule not found with id: " + id))
-        );
+        return assignmentMapper.convertToDTO(assignmentRepository.findById(id).orElseThrow(() -> new ElementNotFoundException("FeedingSchedule not found with id: " + id)));
     }
 
     @Override
     public void delete(Long id) {
+        if (!assignmentRepository.existsById(id)) {
+            throw new EntityNotFoundException("University not found with id: " + id);
+        }
         assignmentRepository.deleteById(id);
+    }
+
+    @Override
+    public void insertFileToAssignment(Long id, File file) {
+        Assignment assignment = assignmentRepository.findById(id)
+                .orElseThrow(() -> new ElementNotFoundException("University not found with id: " + id));
+        assignment.getFiles().add(file);
+        assignmentRepository.save(assignment);
     }
 
     @Override
@@ -63,14 +75,19 @@ public class AssignmentServiceImplementation implements AssignmentService {
 
     @Override
     public Optional<AssignmentDto> update(Long id, AssignmentDto assignmentDto) {
-        return assignmentRepository.findById(id)
-                .map(existingUniversity -> {
-                    Assignment updatedAssignment = assignmentMapper.convertToEntity(assignmentDto);
-                    updatedAssignment.setId(id);
-                    Assignment savedAssignment = assignmentRepository.save(updatedAssignment);
-                    return Optional.of(assignmentMapper.convertToDTO(savedAssignment));
-                })
-                .orElseThrow(() -> new EntityNotFoundException("University not found with id " + id));
+        if (!assignmentRepository.existsById(id)) {
+            throw new ElementNotFoundException("University not found with id: " + id);
+        }
+        assignmentDto.setUpdatedAt(LocalDate.now());
+        Assignment newAssignment = assignmentMapper.convertToEntity(assignmentDto);
+        Assignment assignment = assignmentRepository.findById(id)
+                .orElseThrow(() -> new ElementNotFoundException("University not found with id: " + id));
+        List<File> combinedFiles = new ArrayList<>(newAssignment.getFiles());
+        assignment.getFiles().stream()
+                .filter(file -> !combinedFiles.contains(file))
+                .forEach(combinedFiles::add);
+        newAssignment.setFiles(combinedFiles);
+        return Optional.ofNullable(assignmentMapper.convertToDTO(assignmentRepository.save(newAssignment)));
     }
 
     private <T> List<T> limitListOrDefault(List<T> list, int limit) {
